created FaceInputProvider and TikiModule example
FaceInputProvider in aether.core uses Haar Cascade Face Detection in
OpenCV to find what it thinks is the most likely face in the captured
image.  See docstrings to see how it works.

TikiModule in new examples directory uses FaceInputProvider and puts a
tiki mask over the face it finds!  < 50 lines of code. Our system is so
cool.

Modified driver to accept InputProvider as a constructor keyword arg, I
think we should follow this pattern where the InputProvider is
subclassed and handles all of its own calibration & c.

Improved StubModule slightly to prevent it from crashing