Added first pass for the rewritten Settings class
aether.core.Settings now uses YAML for the file format rather than the
INI style format ConfigParser uses.  This was mostly for the feature of
yaml to automatically determine config value types and create the
appropriate python object.  ConfigParser simply creates strings by
default.  If you want a specific type, you have to know what type to
parse out of the config file.  YAML does this automatically.

A check for the yaml python module was added to the root wscript.

The start of the 'run only this' script was added as bin/aether. This
script will eventually be the only thing we need to run to start aether.
This script intalls to ${PREFIX}/aether/bin/aether.