Added AetherObject.settings_set
AetherObject.settings_set will set a settings value of a given name in a
given section.  This method optionally takes a boolean argument,
test_exist, which if true will only set the settings value if the value
doesn't already exist.

This method will probably be useful when adding feature #61 and fixing
bug #59 and bug #62.