Added AetherObject.settings_load method
AetherObject.settings_load is another helper method for setting up
settings values.  It takes a section name and a dictionary of settings
name/value pairs.  Just as with settings_set, it optionally takes a
boolean (test_exist).  If true, this will only set the values if they do
not already exist.

This will make issue #61 much easier to implement.