Added AetherObject.__setattr__ method
This commit closes #62.  The method AetherObject.__setattr__ was added.
This method was modified to set settings values in addition to object
attributes.

The AetherObject.settings_get method was added and is used in
AetherObject.__getattr__ rather the direct settings access.

The AetherObject.settings_load method was updated to optionally raise
exceptions if trying to set a value that doesn't exist.  Usage of this
method in AetherDriver was updated.